feat: add util fzhist to fuzzy search bash history
It Reads the contents of the bash history file and display them in `fzf`
The selected command will be copied to the clipboard.

Usage:

Run command `fzhist` to read bash history to fzf.

The fzf utility will open a fuzzy search interface with your command history .

Type in a search term to filter by the search term.

Use the arrow keys to navigate the search results and select a command.

Press Enter to execute the selected command.

NOTE: It is similar to binding fzf to Ctrl+R for reverse history search.